+++ src/components/dropdown/dropdown.ts
@@ -65,7 +65,7 @@ export default class SlDropdown extends LitElement {
     | 'left-end' = 'bottom-start';
 
   /** Disables the dropdown so the panel will not open. */
-  @property({ type: Boolean }) disabled = false;
+  @property({ type: Boolean, reflect: true }) disabled = false;
 
   /**
    * By default, the dropdown is closed when an item is selected. This attribute will keep it open instead. Useful for
@@ -408,14 +408,13 @@ export default class SlDropdown extends LitElement {
         flip(),
         shift(),
         size({
-          apply: ({ width, height }) => {
+          apply: ({ availableWidth, availableHeight }) => {
             // Ensure the panel stays within the viewport when we have lots of menu items
             Object.assign(this.panel.style, {
-              maxWidth: `${width}px`,
-              maxHeight: `${height}px`
+              maxWidth: `${availableWidth}px`,
+              maxHeight: `${availableHeight}px`
             });
-          },
-          padding: 8
+          }
         })
       ],
       strategy: this.hoist ? 'fixed' : 'absolute'