+++ src/components/icon/icon.ts
@@ -27,7 +27,7 @@ export default class SlIcon extends LitElement {
   @state() private svg = '';
 
   /** The name of the icon to draw. */
-  @property() name?: string;
+  @property({ reflect: true }) name?: string;
 
   /**
    * An external URL of an SVG file.
@@ -40,7 +40,7 @@ export default class SlIcon extends LitElement {
   @property() label = '';
 
   /** The name of a registered custom icon library. */
-  @property() library = 'default';
+  @property({ reflect: true }) library = 'default';
 
   /** Enforce v-once for vueJs */
   @property({reflect: true, type: Boolean, attribute: 'v-once'}) vueonce = true;